+++ CHANGELOG.md
@@ -17,6 +17,19 @@ y este proyecto utiliza [SemVer](https://semver.org/lang/es/).
 
 ---
 
+## [1.24.5] - 2026-03-18
+
+### Added
+- Added twitch live status sync
+
+### Changed
+- Improved streamer service get live status to use mget instead of multiple get calls
+
+## [1.24.4] - 2026-03-16
+
+### Changed
+- Several backoffice optimizations
+
 ## [1.24.3] - 2026-03-15
 
 ### Changed